change and payement

+++ controller/CategoryController.js
@@ -1,6 +1,8 @@
 
 const AWS = require('aws-sdk')
 require("dotenv").config()
+const jwt = require("jsonwebtoken")
+const { v4: uuidv4 } = require('uuid');
 
 const Category = require("../model/CategoryModule")
 
@@ -15,15 +17,24 @@ const s3 = new AWS.S3({
 const CreateCategory = async (req, res) => {
     const { category_name } = req.body
     const Image = req.file
-    console.log(category_name, Image.originalname)
+    const { authorization } = req.headers
+    const auth = authorization?.split(" ")[1]
 
     try {
+        const { id } = jwt.verify(auth, process.env.SECURE_KEY)
+        console.log(id)
+        if (!id) {
+            return res.status(403).json({
+                message: "you are not authorize"
+            })
+        }
         if (!Image) {
             return res.status(403).json({
                 message: "image is required"
             })
         }
-        const Key = `${Date.now()}-${Image.originalname}`;
+        let newFilename = Image.originalname.replace(/\s+/g, '').replace(/\(1\)/g, '');
+        const Key = `${Date.now()}-${newFilename}`;
 
         const params = {
             Bucket: process.env.AWS_BUCKET_NAME,
@@ -36,11 +47,13 @@ const CreateCategory = async (req, res) => {
 
         const create = await Category.create({
             category_name: category_name,
-            image_url: s3UploadResult.Location
+            image_url: s3UploadResult.Location,
+            category_id: uuidv4(),
+            user_id: id
         })
 
         return res.status(201).json({
-            message: "category",
+            message: "category created successful",
             create
         })
 
@@ -54,8 +67,19 @@ const CreateCategory = async (req, res) => {
 }
 
 const GetCategory = async (req, res) => {
+    const { authorization } = req.headers
+    const auth = authorization?.split(" ")[1]
     try {
-        const get = await Category.findAll({})
+        const { id } = jwt.verify(auth, process.env.SECURE_KEY)
+        
+        if (!id) {
+            return res.status(403).json({
+                message: "you are not authorize"
+            })
+        }
+        const get = await Category.findAll({where:{
+            user_id:id
+        }})
         return res.status(200).json({
             message: "category found",
             data: JSON.stringify(get)
@@ -70,24 +94,43 @@ const GetCategory = async (req, res) => {
 }
 
 const DeleteCategory = async (req, res) => {
-    const {image_path} = req.headers
-    const {id} = req.params
+    const { image_path } = req.headers
+    const { id } = req.params
+    const {authorization} = req.headers
+    const auth = authorization?.split(" ")[1]
     try {
+        const varify =  jwt.verify(auth,process.env.SECURE_KEY)
+        if(!id){
+            return res.status(403).json({
+                message:"you are not authorize"
+            })
+        }
         const image = image_path.split("/");
-        
+
         const params = {
-            Bucket : process.env.AWS_BUCKET_NAME,
-            Key : image[3]
+            Bucket: process.env.AWS_BUCKET_NAME,
+            Key: image[3]
         }
         await s3.deleteObject(params).promise()
-
-       const data =  await Category.destroy({
+        const data1 =   await Category.findAll({
             where: {
-                id:id,
+                user_id:varify.id,
+                category_id:id
             } 
         })
+
+        if(!data1){
+            return res.status(400).json({
+                message:"data not found"
+            })
+        }
+        const data = await Category.destroy({
+            where: {
+                category_id: id,
+            }
+        })
         return res.status(201).json({
-            message:"category deleted ",
+            message: "category deleted ",
             data
         })
     } catch (err) {
@@ -98,9 +141,30 @@ const DeleteCategory = async (req, res) => {
     }
 }
 
+const GetAllCategory = async (req,res) => {
+    try{
+        const get = await Category.findAll({})
+        if(!get){
+            return res.status(403).json({
+                message:"data not found"
+            })
+        }
+        return res.status(200).json({
+            message:"data found",
+            data:JSON.stringify(get)
+        })
+    }catch(err){
+        console.log("get all category error")
+        return res.status(500).json({
+            message:"internal server error"
+        })
+    }
+}
+
 
 module.exports = {
     CreateCategory,
     GetCategory,
-    DeleteCategory
+    DeleteCategory,
+    GetAllCategory
 }
